fix(status color): fix behavior of autosuggesiton and blink
For autosuggestion when is one it showed icons as grey even though it was running, while if
reveresed then does who used blink copilot plugin it will show as if it wasn't running. So took both
in consideration. Thanks Susensio and Eliasnorrby for the issue. Sorry it took me time. Has been
hard this days with time.

fix #15